+++ berkeley-function-call-leaderboard/bfcl/model_handler/local_inference/base_oss_handler.py
@@ -27,9 +27,9 @@ class OSSHandler(BaseHandler, EnforceOverrides):
         self.model_style = ModelStyle.OSSMODEL
         self.dtype = dtype
 
-        # Might be overridden in batch_inference method if local_model_path is provided
+        # Will be overridden in batch_inference method
         # Used to indicate where the tokenizer and config should be loaded from
-        self.model_path_or_id = self.model_name_huggingface
+        self.model_path_or_id = None
 
         # Read from env vars with fallbacks
         self.vllm_host = os.getenv("VLLM_ENDPOINT", "localhost")
@@ -99,6 +99,7 @@ class OSSHandler(BaseHandler, EnforceOverrides):
                 "trust_remote_code": True,
             }
         else:
+            self.model_path_or_id = self.model_name_huggingface
             load_kwargs = {
                 "pretrained_model_name_or_path": self.model_path_or_id,
                 "trust_remote_code": True,